Refactor: 페이지 이동을 위하여 `Insight` 분리 -> 목록, 상세 정보
